fix(desktop): sign-extend hwnd pointer in 64-bit windows window validation
Casting the u32 window id straight to usize zero-extends it on 64-bit
Windows. Handles with the top bit set (0x8xxxxxxx) became malformed
pointers, IsWindow returned 0 and the capture died on the first frame
claiming the window had been closed. Going through i32 as isize
sign-extends the handle the way Win32 expects.

+++ web/src-tauri/src/screen_capture.rs
@@ -692,7 +692,8 @@ fn is_window_valid(window_id: u32) -> bool {
         fn IsWindow(hwnd: *mut std::ffi::c_void) -> i32;
     }
     // SAFETY: IsWindow receives an HWND pointer-sized value and returns 0 if invalid.
-    unsafe { IsWindow(window_id as usize as *mut std::ffi::c_void) != 0 }
+    // Em Windows 64-bit, handles HWND de 32-bit precisam de sign-extension (i32 -> isize).
+    unsafe { IsWindow(window_id as i32 as isize as *mut std::ffi::c_void) != 0 }
 }
 
 #[cfg(not(windows))]
